requirements.txt
mysql-connector-python>=8.0.31
python-decouple>=3.4
requests>=2.26.0
flask>=2.2.2
python-dotenv>=0.21.0